--- a/encoder.go
+++ b/encoder.go
@@ -83,7 +83,7 @@ func (e encoder) defaultEncode(num uint128) string { // compiler optimizes a lot
 func (e encoder) encode(num uint128) string {
 	var r, ind uint64
 	i := e.alphabet.encLen - 1
-	buf := make([]byte, e.alphabet.encLen*e.alphabet.maxBytes)
+	buf := make([]byte, e.alphabet.encLen*int64(e.alphabet.maxBytes))
 	curByteInd := len(buf) - 1
 	l := uint64(e.alphabet.len)
 	d, n := maxPow(l)