Key changes:
 - simplified encoder complexity (now only 2 cases instead of 4, one for default alphabet, other for generic)
 - changed custom alphabets encode to efficiently place runes into a byte slice backwards and avoid unnecessary allocations
 - added `hashedUUID` function to reduce allocation in `NewWithNamespace`, as google's `uuid.NewSHA1` makes 4 unnecessary allocations

Benchamarks:
```
goos: linux
goarch: amd64
pkg: github.com/lithammer/shortuuid/v4
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
                         │   v4.2.0    │                 new                 │
                         │   sec/op    │   sec/op     vs base                │
UUID-12                    123.3n ± 0%   119.6n ± 0%   -3.04% (p=0.000 n=10)
Encoding-12                44.42n ± 1%   42.32n ± 0%   -4.74% (p=0.001 n=10)
EncodingB57_MB-12          123.2n ± 4%   116.5n ± 2%   -5.36% (p=0.001 n=10)
EncodingB16-12             134.5n ± 0%   151.9n ± 0%  +12.98% (p=0.000 n=10)
EncodingB16_MB-12          327.0n ± 1%   191.2n ± 2%  -41.53% (p=0.000 n=10)
Decoding-12                173.9n ± 1%   167.4n ± 1%   -3.79% (p=0.000 n=10)
DecodingB16-12             206.0n ± 1%   200.0n ± 1%   -2.91% (p=0.000 n=10)
DecodingB16_MB-12          254.6n ± 1%   249.9n ± 2%   -1.87% (p=0.015 n=10)
NewWithAlphabet-12         387.2n ± 1%   381.0n ± 2%   -1.61% (p=0.008 n=10)
NewWithAlphabetB16-12      270.4n ± 1%   293.6n ± 1%   +8.56% (p=0.000 n=10)
NewWithAlphabetB16_MB-12   553.5n ± 0%   403.5n ± 1%  -27.10% (p=0.000 n=10)
NewWithNamespace-12        249.3n ± 1%   166.3n ± 1%  -33.29% (p=0.000 n=10)
NewWithNamespaceHttp-12    255.1n ± 3%   169.8n ± 1%  -33.43% (p=0.000 n=10)
NewWithNamespaceHttps-12   256.4n ± 2%   167.4n ± 1%  -34.73% (p=0.000 n=10)
geomean                    206.5n        177.4n       -14.08%

                         │    v4.2.0     │                 new                  │
                         │     B/op      │    B/op     vs base                  │
UUID-12                     24.00 ± 0%     24.00 ± 0%        ~ (p=1.000 n=10) ¹
Encoding-12                 24.00 ± 0%     24.00 ± 0%        ~ (p=1.000 n=10) ¹
EncodingB57_MB-12           72.00 ± 0%     80.00 ± 0%  +11.11% (p=0.000 n=10)
EncodingB16-12              64.00 ± 0%     32.00 ± 0%  -50.00% (p=0.000 n=10)
EncodingB16_MB-12          480.00 ± 0%     96.00 ± 0%  -80.00% (p=0.000 n=10)
Decoding-12                 0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
DecodingB16-12              0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
DecodingB16_MB-12           0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewWithAlphabet-12          264.0 ± 0%     264.0 ± 0%        ~ (p=1.000 n=10) ¹
NewWithAlphabetB16-12      128.00 ± 0%     96.00 ± 0%  -25.00% (p=0.000 n=10)
NewWithAlphabetB16_MB-12    544.0 ± 0%     160.0 ± 0%  -70.59% (p=0.000 n=10)
NewWithNamespace-12        200.00 ± 0%     24.00 ± 0%  -88.00% (p=0.000 n=10)
NewWithNamespaceHttp-12    208.00 ± 0%     24.00 ± 0%  -88.46% (p=0.000 n=10)
NewWithNamespaceHttps-12   224.00 ± 0%     24.00 ± 0%  -89.29% (p=0.000 n=10)
geomean                                ²               -51.82%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                         │    v4.2.0    │                 new                  │
                         │  allocs/op   │ allocs/op   vs base                  │
UUID-12                    1.000 ± 0%     1.000 ± 0%        ~ (p=1.000 n=10) ¹
Encoding-12                1.000 ± 0%     1.000 ± 0%        ~ (p=1.000 n=10) ¹
EncodingB57_MB-12          2.000 ± 0%     1.000 ± 0%  -50.00% (p=0.000 n=10)
EncodingB16-12             2.000 ± 0%     1.000 ± 0%  -50.00% (p=0.000 n=10)
EncodingB16_MB-12          4.000 ± 0%     1.000 ± 0%  -75.00% (p=0.000 n=10)
Decoding-12                0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
DecodingB16-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
DecodingB16_MB-12          0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewWithAlphabet-12         2.000 ± 0%     2.000 ± 0%        ~ (p=1.000 n=10) ¹
NewWithAlphabetB16-12      3.000 ± 0%     2.000 ± 0%  -33.33% (p=0.000 n=10)
NewWithAlphabetB16_MB-12   5.000 ± 0%     2.000 ± 0%  -60.00% (p=0.000 n=10)
NewWithNamespace-12        5.000 ± 0%     1.000 ± 0%  -80.00% (p=0.000 n=10)
NewWithNamespaceHttp-12    5.000 ± 0%     1.000 ± 0%  -80.00% (p=0.000 n=10)
NewWithNamespaceHttps-12   5.000 ± 0%     1.000 ± 0%  -80.00% (p=0.000 n=10)
geomean                               ²               -47.13%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```